Reorganizing and resolving warnings
Splits EMATCH main file to multiple files, and resolves warnings
