+++ src/Makefile
@@ -211,6 +211,9 @@ core_src_c := \
 	solvers/simplex/simplex.c \
 	solvers/quant/ef_problem.c \
 	solvers/quant/quant_pattern.c \
+	solvers/quant/ematch_instr.c \
+	solvers/quant/ematch_instr_stack.c \
+	solvers/quant/ematch_compile.c \
 	solvers/quant/quant_ematching.c \
 	solvers/quant/quant_solver.c \
 	terms/balanced_arith_buffers.c \
